Switch to Alembic-driven migrations and fix Deezer ID overflow
- Replace create_all in lifespan with entrypoint that runs alembic upgrade head
- Add entrypoint.sh executing migrations before app start
- Change deezer_id from Integer to BigInteger in entities and migrations
- Add search/getArtist/getAlbum/getSong routes without .view suffix
  for frontend compatibility

+++ rivo_drome/entity/album.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy import BigInteger, Column, Integer, String, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 
 from rivo_drome.entity.base import BaseEntity
@@ -13,7 +13,7 @@ class Album(BaseEntity):
     title = Column(String(255), nullable=False, index=True)
     artist_id = Column(Integer, ForeignKey("artist.id"), nullable=False)
     cover_url = Column(String(512), nullable=True)
-    deezer_id = Column(Integer, unique=True, nullable=True)
+    deezer_id = Column(BigInteger, unique=True, nullable=True)
     year = Column(Integer, nullable=True)
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False)
